Re-structuring and documentation edits
Network functions are now free functions, due to the nature of
projection / transformation operations: network download and endpoint
functionality must be set prior to instantiating a Proj instance, so
it doesn't make any sense to implement these functions on Proj itself.

The only drawback is that the settings are global, and apply to all
subsequent proj instances. However, this is unlikely to be a problem
in practice.