Re-write HTPP error-handling
S3 has some retry error codes that are officially client errors. The
new logic for inital reads and content-range reads checks whether a
response is a server error or one of these "special" error codes, and
executes the retry logic if so. Otherwise (or subsequently), any
error is checked for, and the functions return early.