rename Point trait -> Coord
It more closely corresponds with how we're using the terminology in geo,
and paves the way for interacting with "a bunch of coords" from
geo-types

e.g. if we're projecting a linestring, we'd operate on the linestrings
coords, not it's points.

Not a big deal, but since we're introducing a new thing (the trait), no
reason to inherit the old name if another fits better.