Fix source build on aarch64/homebrew
On apple platforms, libtiff is not installed by the operating system. If
the user has it installed, likely it was installed by homebrew.

Previously, on x86, homebrew installed libs into /usr/lib which is in
the default search path, but since aarch64, homebrew has moved libs to
/opt/homebrew/lib.

So now we use pkg-config to find the library.

Potential problems:

 - Not everyone has pkg-config (e.g windows users, a few mac users)
 - Not everyone actually needs libtiff, only the network geotiff users,
   but we assume everyone needs it since most system installations
   include it.

Note that proj enables tiff support by defeault, provided libtiff is
found (using pkg-config? or something else?)