Don't pass RequestBuilder around in an opaque pointer
The RequestBuilder struct is complex and uses Tokio under the hood.
Instead of passing it around, we pass the URL around and create a new
Client for data retrieval. This is a bit less efficient since we don't
use the connection pool created by the initial Client, but it's also
safer.