duplicate method impls into Proj/ProjBuilder so user doesn't need to `use proj::Info` to have access to these methods.
This is a nice convenience, but especially important because we want to
have `use proj::Info` refer to the newly renamed `Info` struct, not the
legacy `Info` trait.

So now, instead of legacy people getting a confusing method for trying
to use the legacy trait, they'll just see an "unused import" warning.